Extract lineage.js from gameLogic.js (closes #119)
