+++ src/screens/GlobalCombatantDetail.jsx
@@ -4,7 +4,8 @@ import TagChips from '../components/TagChips.jsx'
 import TagInput from '../components/TagInput.jsx'
 import { btn, inp, lbl } from '../styles.js'
 import { updateGlobalCombatant, getLineageTree, getCombatantRoundHistory, sget, superHostSetEntityTags, superHostInductHoF, superHostRemoveHoF, superHostEditHofNote, setCombatantGroups, getCombatantGroupIds, listPublishedGroups, getAwardsForCombatant } from '../supabase.js'
-import { buildStoryFromLineageTree, computeSuperlatives, AWARD_TYPE_LABELS } from '../gameLogic.js'
+import { computeSuperlatives, AWARD_TYPE_LABELS } from '../gameLogic.js'
+import { buildStoryFromLineageTree } from '../lineage.js'
 import { downloadFile, formatCombatantHistory } from '../export.js'
 import GameSummaryScreen from './GameSummaryScreen.jsx'
